[Copy] Available Storage Check
This approach makes one assumption:
   (block size of install path) == (block size of unpacked path)
In the case that this is not true, it will work incorrectly.

This approach will report OUT_OF_SPACE error in two cases:
 1) directory or file cannot be unzipped.
 2) unpacked directory is too large to be moved to installation location.

There is no possibility to calculate size of directory entires
to be created to unzip widget so installer just needs to unzip it.

=== 'COMMIT DISCLAIMER' ===

Note that size of directory entry depends on:
 - internal representation of directory for given fs type
 - number of entries in directory,

This methods DOESN'T calculate space for:
 - icons for platform use,
 - xml platfom manifest,
 - binary symlink in application directory.

Any suggestions, welcome...

===========================

Change-Id: Ib8e81c4cb56e771347a128021ece53e4cf2c9fce