Reimplement signature validation with cert-svc/vcore
This commit does not introduce any new feature in app-installer.
Signature is reimplemented only. Prililege level check and
other check will be implemented afterwards.

Removed dependencies to openssl and xmlsec as we are using
cert-svc now.

Reworked basic tests about signature:
  "signature_unittest.cc"

Removed LICENSE-xwalk as it was not used anymore. All crosswalk
code was removed/moved to manifest-parser or rewritten.

Change-Id: I59462fd0e0082f5c2ec648caedebba8874e69ddf