Fix build for g++ without optimalization flags
The problem is that value static field was supposed to be inlined
but it is not. It appears to work only for optimalization flags
 with g++.

Full explanation link:
 - http://stackoverflow.com/questions/8016780/undefined-reference-to-static-constexpr-char

In following code value field appears to be "odr-used" in context
of ostream operator<<. Conversion lvalue-to-rvalue should be
"immediately applied" in LOG marco. Therefore this commit adds
construction of std::string which has implicit constructor of const char*.

Change-Id: I2f8c9c57b14d9f7c6a7581dd5c8aab07f699d04b