Revoke security rules after removing files on uninstall
On uninstall steps, old code runs StepRevokeSecurity before
StepRemoveFiles. With SMACK enabled, once StepRevokeSecurity is
finished, the installer processes(tpk and wgt) don't have an authority
to remove files, so the app directory($HOME/apps_rw/[pkgid]) is still
remained. When the same app is installed again, the installer process
crashes, as the boost::exists() to the app destination directory raises
an exception on SMACK denial.
This patch moves the StepRevokeSecurity to the end of the list of steps
on uninstall, to give the privilege to delete the app directory to the
installers.

Change-Id: I17cc9ea6170352242eede2502458a38944e82826
Signed-off-by: Youmin Ha <youmin.ha@samsung.com>