Fix deinstallation mode with smack
As stated in: https://review.tizen.org/gerrit/#/c/46420/

On uninstall steps, old code runs StepRevokeSecurity before
StepRemoveFiles. With SMACK enabled, once StepRevokeSecurity
is finished, the installer processes(tpk and wgt) don't have
an authority to remove files, so the app
directory($HOME/apps_rw/[pkgid]) is still remained. When the
same app is installed again, the installer process crashes,
as the boost::exists() to the app destination directory raises
an exception on SMACK denial. This patch moves the
StepRevokeSecurity to the end of the list of steps on uninstall,
to give the privilege to delete the app directory to the installers.

Splitting StepRevokeSecurity into two parts so that deinstallation
and deinstallation rollback should work.

Change-Id: If597a3c256e09d57aa923566f5c3454d497c0593