Add pkgdir_maker tool for global apps
Global apps are installed in /usr/apps/, but this path is not allowed for
users to write their data.  To store app data of global apps, the data and
cache directories have to be created.

This tool does two things:
 1. creates the global app package's data and cache directories
   ({pkgid}/data, {pkgid}/cache) into all user's apps_rw/ directories.
   Proper owner, group, permission and SMACK labels are set for those dirs.
 2. For new users in the future, the same directories are create into the
   user home template, /etc/skel/.

NOTE: For prototyping, actual implementation is made as a bash script
(pkgdir_maker_impl.sh), but later it is to be replaced with native code.

Change-Id: I8ca17362a2c33ebe4a051f80a96d9bb4263dabfa
Signed-off-by: Youmin Ha <youmin.ha@samsung.com>