Add new step for adding default privileges
Add some privileges as default in next cases.
If the package is wgt,
  add 'http://tizen.org/privilege/internal/webappdefault'.
If the package has public cert,
  add 'http://tizen.org/privilege/internal/default/public'.
If the package has partner cert,
  add 'http://tizen.org/privilege/internal/default/partner'.
If the package has platform cert,
  add 'http://tizen.org/privilege/internal/default/platform'.
These default privileges will be converted to mapped privileges by
the security-manager.

NOTE: api version 3.0 is not considered yet. should be fixed later.

Change-Id: I70d0ec09efbcd8c80e6189412c8bdbac84ef438a