Add PluginsLauncher and DynamicLibHandle
This change adds the class PluginsLauncher which is to be used by
StepRunParserPlugin. The DynamicLibHandle class is used by PluginsLauncher
to manage plugins which are in the form of dynamic libraries.
This commit can be submitted independently, however to launch plugins
(via PluginLauncher class) StepRunParserPlugin needs to be submitted too.

Change-Id: I55daa5f8d8867e4716530db5d6c9319526bf02fe