Add methods for testing in PkgMgrInterface
Smoke tests are randomly crashing since some changes with dbus
connection closing in slp-pkgmgr. It seems to be some time hazard (?)
on closing connection.

Anyway, tests are setting up and closing connection repeatingly.
To avoid this, for testing, CreateForTesting() will be used
to avoid crash. This will allow to run all smoke tests in one process
(not one by one).

Change-Id: I69f5972c09a593cfe9a282f2120845f85aa65b52