set 'rpm' type to rpm packed application packages
to support the old behaviors use 'type' information(like as 'rpm') for their own purpose.
we can't not figure out which modules use 'type' info in previous entire models and,
it not easy to make them fix the behavior to other way.
so, now we realive the 'rpm' type.
only rpm packed application package can be 'rpm' type.

Change-Id: I7c7e40e5fc925e1daff5629a36e7ba47acddb8e7
Signed-off-by: jongmyeongko <jongmyeong.ko@samsung.com>