Removing proceeding account and shortcut data
This data is parsed by plugins in case of tpk and we don't
need to store this information anywhere. Information that
was set by those methods wasn't read anywhere.

There is a different situation in case of wgt backend where
we generate platform xml manifest and we need this information.
Otherwise it may be lost when update is running.

Change-Id: I73a073e0ff252b9cef7445143965d708d7a2c5d4